backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
tensorflow==2.13.0
keras==2.13.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
yfinance==0.2.18
psycopg2-binary==2.9.7
python-dotenv==1.0.0
werkzeug==2.3.7
